src/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
av==15.0.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
elevenlabs==2.8.0
fastapi==0.116.1
google-cloud-texttospeech==2.27.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
numpy==2.3.2
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
scipy==1.16.1
sniffio==1.3.1
speechify-api==1.2.0
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
